request permission smarter
