+++ app/src/main/java/black/bracken/neji/ui/scanqrcode/ScanQrCodeFragment.kt
@@ -1,28 +1,17 @@
 package black.bracken.neji.ui.scanqrcode
 
-import android.Manifest
-import android.annotation.SuppressLint
-import android.content.pm.PackageManager
 import android.os.Bundle
 import android.view.View
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.camera.core.CameraSelector
 import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageProxy
 import androidx.camera.core.Preview
 import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import black.bracken.neji.R
 import black.bracken.neji.databinding.ScanQrCodeFragmentBinding
-import com.google.android.material.snackbar.Snackbar
-import com.google.mlkit.vision.barcode.Barcode
-import com.google.mlkit.vision.barcode.BarcodeScannerOptions
-import com.google.mlkit.vision.barcode.BarcodeScanning
-import com.google.mlkit.vision.common.InputImage
 import com.wada811.viewbinding.viewBinding
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.concurrent.ExecutorService
@@ -31,9 +20,7 @@ import java.util.concurrent.Executors
 @AndroidEntryPoint
 class ScanQrCodeFragment : Fragment(R.layout.scan_qr_code_fragment) {
 
-    private val viewModel: ScanQrCodeViewModel by viewModels()
     private val binding by viewBinding(ScanQrCodeFragmentBinding::bind)
-
     private val args: ScanQrCodeFragmentArgs by navArgs()
 
     private val cameraExecutor: ExecutorService by lazy { Executors.newSingleThreadExecutor() }
@@ -41,22 +28,7 @@ class ScanQrCodeFragment : Fragment(R.layout.scan_qr_code_fragment) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        if (allPermissionsGranted()) {
-            startCamera()
-        } else {
-            registerForActivityResult(ActivityResultContracts.RequestPermission()) { isAllowed ->
-                if (isAllowed) {
-                    startCamera()
-                } else {
-                    // TODO: handle errors
-                    Snackbar
-                        .make(view.rootView, "you have no permission", Snackbar.LENGTH_SHORT)
-                        .show()
-                }
-            }
-        }
-
-        cameraExecutor // initialize
+        startCamera()
     }
 
     override fun onDestroy() {
@@ -94,7 +66,6 @@ class ScanQrCodeFragment : Fragment(R.layout.scan_qr_code_fragment) {
 
             try {
                 cameraProvider.unbindAll()
-
                 cameraProvider.bindToLifecycle(this, cameraSelector, preview, imageAnalyzer)
             } catch (ex: Exception) {
                 // TODO: handle errors
@@ -102,40 +73,6 @@ class ScanQrCodeFragment : Fragment(R.layout.scan_qr_code_fragment) {
             }
 
         }, ContextCompat.getMainExecutor(requireContext()))
-    }
-
-    private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all { permission ->
-        requireContext().checkSelfPermission(permission) == PackageManager.PERMISSION_GRANTED
-    }
-
-    companion object {
-        private const val REQUEST_CODE_PERMISSIONS = 195
-        private val REQUIRED_PERMISSIONS = listOf(Manifest.permission.CAMERA)
-    }
-
-    private class QrCodeAnalyzer(private val onScan: (String) -> Unit) : ImageAnalysis.Analyzer {
-
-        private val scanner = BarcodeScanning.getClient(
-            BarcodeScannerOptions.Builder()
-                .setBarcodeFormats(Barcode.FORMAT_QR_CODE)
-                .build()
-        )
-
-        @SuppressLint("UnsafeExperimentalUsageError")
-        override fun analyze(proxy: ImageProxy) {
-            val media = proxy.image ?: return
-            val image = InputImage.fromMediaImage(media, proxy.imageInfo.rotationDegrees)
-
-            scanner.process(image)
-                .addOnSuccessListener { qrCodes ->
-                    qrCodes
-                        .firstOrNull()
-                        ?.rawValue
-                        ?.also { text -> onScan(text) }
-                }
-                .addOnCompleteListener { proxy.close() }
-        }
-
     }
 
 }
